Address many review comments and fix documentation
